requirements.txt
websocket-client==1.6.1
python-dotenv==1.0.0
